iter15: Изменен пакет storage
