+++ internal/handlers/handlers.go
@@ -6,8 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/JuliyaMS/service-metrics-alerting/internal/config"
-	"github.com/JuliyaMS/service-metrics-alerting/internal/database"
-	"github.com/JuliyaMS/service-metrics-alerting/internal/file"
 	"github.com/JuliyaMS/service-metrics-alerting/internal/hash"
 	"github.com/JuliyaMS/service-metrics-alerting/internal/html"
 	"github.com/JuliyaMS/service-metrics-alerting/internal/logger"
@@ -23,12 +21,13 @@ import (
 )
 
 type Handlers struct {
-	memStor storage.Repositories
+	MemStor storage.Repositories
 }
 
 func NewHandlers(stor storage.Repositories) *Handlers {
+	fmt.Println(stor)
 	return &Handlers{
-		memStor: stor,
+		MemStor: stor,
 	}
 }
 
@@ -43,7 +42,7 @@ func (h *Handlers) requestValue(w http.ResponseWriter, r *http.Request) {
 	metricName := chi.URLParam(r, "name")
 	metricValue := chi.URLParam(r, "value")
 
-	if err := h.memStor.Add(metricType, metricName, metricValue); err != nil {
+	if err := h.MemStor.Add(metricType, metricName, metricValue); err != nil {
 		w.WriteHeader(http.StatusBadRequest)
 		return
 	}
@@ -79,8 +78,8 @@ func (h *Handlers) requestGetName(w http.ResponseWriter, r *http.Request) {
 	metricType := chi.URLParam(r, "type")
 	metricName := chi.URLParam(r, "name")
 
-	logger.Logger.Info("Get value from memStor")
-	value := h.memStor.Get(metricType, metricName)
+	logger.Logger.Info("Get value from MemStor")
+	value := h.MemStor.Get(metricType, metricName)
 
 	if value != "-1" {
 		logger.Logger.Info("Get value: ", value, " for metric: ", metricName)
@@ -102,7 +101,7 @@ func (h *Handlers) requestGetAll(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	logger.Logger.Infow("Get data for HTML format")
-	dataGauge, dataCounter := html.GetHTMLStructs(h.memStor.GetAll())
+	dataGauge, dataCounter := html.GetHTMLStructs(h.MemStor.GetAll())
 
 	w.Header().Set("Content-Type", "text/html")
 	w.WriteHeader(http.StatusOK)
@@ -172,7 +171,7 @@ func (h *Handlers) requestUpdate(w http.ResponseWriter, r *http.Request) {
 	}
 	if req.MType == "gauge" {
 		logger.Logger.Infow("Update gauge metric", "name", req.ID, "value", *req.Value)
-		if err := h.memStor.Add(req.MType, req.ID, strconv.FormatFloat(*req.Value, 'g', -1, 64)); err != nil {
+		if err := h.MemStor.Add(req.MType, req.ID, strconv.FormatFloat(*req.Value, 'g', -1, 64)); err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			return
 		}
@@ -180,12 +179,12 @@ func (h *Handlers) requestUpdate(w http.ResponseWriter, r *http.Request) {
 
 	if req.MType == "counter" {
 		logger.Logger.Infow("Update counter metric", "name", req.ID, "value", *req.Delta)
-		if err := h.memStor.Add(req.MType, req.ID, strconv.FormatInt(*req.Delta, 10)); err != nil {
+		if err := h.MemStor.Add(req.MType, req.ID, strconv.FormatInt(*req.Delta, 10)); err != nil {
 			w.WriteHeader(http.StatusBadRequest)
 			return
 
 		}
-		newDelta, err := strconv.ParseInt(h.memStor.Get("counter", req.ID), 10, 64)
+		newDelta, err := strconv.ParseInt(h.MemStor.Get("counter", req.ID), 10, 64)
 		if err != nil {
 			logger.Logger.Error("cannot write new Delta", zap.Error(err))
 			w.WriteHeader(http.StatusInternalServerError)
@@ -238,9 +237,11 @@ func (h *Handlers) requestGetValue(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
+	fmt.Println(h.MemStor)
 	if req.MType == "gauge" {
 		logger.Logger.Infow("Get gauge metric value", "name", req.ID)
-		Value, err := strconv.ParseFloat(h.memStor.Get(req.MType, req.ID), 64)
+		Value, err := strconv.ParseFloat(h.MemStor.Get(req.MType, req.ID), 64)
+		fmt.Println("Gauge Value:", Value)
 		if err != nil {
 			logger.Logger.Error("cannot write Value", zap.Error(err))
 			w.WriteHeader(http.StatusInternalServerError)
@@ -249,12 +250,14 @@ func (h *Handlers) requestGetValue(w http.ResponseWriter, r *http.Request) {
 		if Value == -1 {
 			Value = 0
 		}
+		fmt.Println("Gauge Value:", Value)
 		req.Value = &Value
 	}
 
 	if req.MType == "counter" {
 		logger.Logger.Infow("Get counter metric value", "name", req.ID)
-		Delta, err := strconv.ParseInt(h.memStor.Get(req.MType, req.ID), 10, 64)
+		Delta, err := strconv.ParseInt(h.MemStor.Get(req.MType, req.ID), 10, 64)
+		fmt.Println("Counter Value:", Delta)
 		if err != nil {
 			logger.Logger.Error("Cannot write Delta", zap.Error(err))
 			w.WriteHeader(http.StatusInternalServerError)
@@ -263,6 +266,7 @@ func (h *Handlers) requestGetValue(w http.ResponseWriter, r *http.Request) {
 		if Delta == -1 {
 			Delta = 0
 		}
+		fmt.Println("Counter Value:", Delta)
 		req.Delta = &Delta
 	}
 
@@ -294,7 +298,7 @@ func (h *Handlers) requestEmpty(w http.ResponseWriter, r *http.Request) {
 func (h *Handlers) PingDB(w http.ResponseWriter, r *http.Request) {
 	logger.Logger.Info("start handler: PingDB")
 
-	if err := h.memStor.CheckConnection(); err != nil {
+	if err := h.MemStor.CheckConnection(); err != nil {
 		logger.Logger.Error("get error while check connection to Database:", err)
 		w.WriteHeader(http.StatusInternalServerError)
 		return
@@ -331,7 +335,7 @@ func (h *Handlers) UpdatesDB(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err := h.memStor.AddAnyData(req); err != nil {
+	if err := h.MemStor.AddAnyData(req); err != nil {
 		logger.Logger.Error("Get error while execute transaction:", err)
 		w.WriteHeader(http.StatusInternalServerError)
 		return
@@ -397,23 +401,16 @@ func routeGet(r *chi.Mux, h *Handlers) {
 	})
 }
 
-func NewRouter(DBConn *database.ConnectionDB) *chi.Mux {
+func NewRouter() (*chi.Mux, *Handlers) {
 	logger.Logger.Infow("init router and handlers")
-	if config.Restore && config.FileStoragePath != "" {
-		logger.Logger.Info("restore data from file:", config.FileStoragePath)
-		err := file.ReadFromFile(config.FileStoragePath)
-		if err != nil {
-			logger.Logger.Errorf(err.Error(), "can't read data from file:", config.FileStoragePath)
-		}
-	}
 
-	var handlers *Handlers
-	if config.DatabaseDsn != "" {
-		handlers = NewHandlers(DBConn)
-	} else {
-		handlers = NewHandlers(&storage.Storage)
+	stor := storage.NewStorage()
+	if stor == nil {
+		logger.Logger.Error("Can't create storage")
+		return nil, nil
 	}
-	handlers.memStor.Init()
+	var handlers = NewHandlers(stor)
+	handlers.MemStor.Init()
 
 	logger.Logger.Info("create new router")
 	r := chi.NewRouter()
@@ -426,5 +423,5 @@ func NewRouter(DBConn *database.ConnectionDB) *chi.Mux {
 	r.Get("/", m.LoggingServer(m.CompressionGzip(handlers.requestGetAll)))
 	r.Get("/ping", m.LoggingServer(handlers.PingDB))
 
-	return r
+	return r, handlers
 }
